+++ modules/controllers/pool-controller.ts
@@ -10,7 +10,7 @@ import { getViemClient } from '../sources/viem-client';
 import { getBlockNumbersSubgraphClient, getV3JoinedSubgraphClient, getVaultSubgraphClient } from '../sources/subgraphs';
 import { prisma } from '../../prisma/prisma-client';
 import { updateLiquidity24hAgo, updateLiquidityValuesForPools } from '../actions/pool/update-liquidity';
-import { updateLifetimeValues } from '../actions/pool/update-holders-count';
+import { updateLifetimeValues } from '../actions/pool/update-liftetime-values';
 import { Chain, PrismaLastBlockSyncedCategory } from '@prisma/client';
 import { getVaultClient } from '../sources/contracts/v3/vault-client';
 import { upsertPools as upsertPoolsV3 } from '../actions/pool/v3/upsert-pools';
